Refactoring code #JC19-326
