+++ apps/vanilla/src/pages/index.html
@@ -47,7 +47,6 @@
             <select name="fields" class="form__select sort-field" id="sort-field"></select>
           </div>
 
-
           <div class="field">
             <label class="form__label field" for="sort-ordering">
               <span class="form__label-text">Ordering:</span>
@@ -61,7 +60,6 @@
             </label>
             <select name="status" class="form__select sort-status" id="sort-status"></select>
           </div>
-
 
         </form>
